+++ src/Structs.jl
@@ -252,22 +252,30 @@ end
 abstract type Shape end
 
 
-
 """
-A point light (used by the point-light renderer)
-This class holds information about a point light (a Dirac's delta in the rendering equation). 
+    PointLight(p::Point, c::Color, r::Float64 = 0.0)
 
-#Arguments
--   `position`: a `Point` object holding the position of the point light in 3D space
--   `color`: the color of the point light (an instance of `RGB{Float32}`)
--   `linear_radius`: a floating-point number. If non-zero, this «linear radius» `r` is used to compute the solid
-    angle subtended by the light at a given distance `d` through the formula `(r / d)²`.
+A point light (used by the point-light renderer).
+This class holds information about a point light (a Dirac's delta in the 
+rendering equation).
+
+## Arguments
+
+- `position::Point` : position of the point light in 3D space
+
+- `color::RGB{Float32}` : color of the point light
+
+- `linear_radius::Float64`: if non-zero, this «linear radius» `r` is 
+  used to compute the solid angle subtended by the light at a given 
+  distance `d` through the formula ``(r / d)^2``.
+
+See also: [`Point`](@ref), [`PointLightRenderer`](@ref)
 """
 struct PointLight
     position::Point
     color::RGB{Float32}
     linear_radius::Float64
-    PointLightRenderer(p, c, r = 0.0) = new(p, c, r)
+    PointLight(p::Point, c::Color, r::Float64 = 0.0) = new(p, c, r)
 end
 
 ##########################################################################################92
@@ -512,7 +520,7 @@ struct PathTracer <: Renderer
     num_of_rays::Int64
     max_depth::Int64
     russian_roulette_limit::Int64
-    PathTracer(w, bc=BLACK, pcg=PCG(), n=10, md=2, RRlim=3) = 
+    PathTracer(w::World, bc=BLACK, pcg=PCG(), n=10, md=2, RRlim=3) = 
         new(w, bc, pcg, n, md, RRlim)
 end
 
@@ -571,4 +579,35 @@ struct PointLightRenderer
     background_color::RGB{Float64}
     ambient_color::RGB{Float64}
     PointLightRenderer(e, bc = RGB(0., 0., 0.), ac = (0.1, 0.1, 0.1)) = new(w, bc, ac)
+end
+
+"""
+    PointLightRenderer(
+        world::World
+        background_color::RGB{Float32} = RGB{Float32}(0., 0., 0.)
+        ambient_color::RGB{Float32} = RGB{Float32}(0.1, 0.1, 0.1)
+    )
+
+A simple point-light tracing renderer.
+
+## Arguments
+
+- `world::World` : the world to be rendered
+
+- `background_color::RGB{Float32}` : default background color 
+  if the Ray doesn-t hit anything
+
+- `ambient_color::RGB{Float32}` : default ambient color 
+
+See also: [`World`](@ref)
+"""
+struct PointLightRenderer <: Renderer
+    world::World
+    background_color::RGB{Float32}
+    ambient_color::RGB{Float32}
+    PointLightRenderer(
+            w::World, 
+            bc=RGB{Float32}(0., 0., 0.), 
+            ac=RGB{Float32}(0.1, 0.1, 0.1)
+        ) = new(w, bc, ac)
 end