+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 <<<<<<< HEAD
 - Added antialiasing algorithm
 
+
+- Added point-light tracing algorithm, see PR[#17](https://github.com/cosmofico97/Raytracing/pull/17)
+
 # VERSION 0.3.0
 
 - Added antialiasing algorithm, see PR[#16](https://github.com/cosmofico97/Raytracing/pull/16)
@@ -10,7 +13,10 @@
 >>>>>>> antialiasing
 
 <<<<<<< HEAD
+<<<<<<< HEAD
 =======
+=======
+>>>>>>> ddb98273ccf586f2bf52ed0b928d7a4851a59b77
 
 # VERSION 0.2.0
 >>>>>>> pointlighttracing